requirements.txt
virtualenv
jupyter
ipykernel

pandas
streamlit

transformers
sentence_transformers
transformers
tensor2tensor
tempun

umap-learn
scikit-learn